all steps works

+++ analyzer/process.py
@@ -13,6 +13,7 @@ from yolo_detection import detect_objects, sort_bib_numbers, extract_bib_numbers
 from model import  generate_json_person
 from color_detection import color_or_grayscale
 from color_classifier import  determine_color_v2
+import json
 
 
 
@@ -29,9 +30,12 @@ def detect_and_segment(img, text_prompt,b_t_dino=0.3, t_t_dino=0.3):
     for phrase in phrases:
         if phrase not in text_prompt:
             LOGGER.warning("La phrase : "+phrase+" n'est pas dans le text_prompt")
-            print("La phrase : "+phrase+" n'est pas dans le text_prompt")
     
     return annotated_image, segmented_image, detections, phrases
+
+def save_img (img, name):
+    cv2.imwrite(name, img)
+
 
 def first_step(img):
     text_prompt = ['person']
@@ -49,13 +53,15 @@ def second_step(colorize_list_of_masks, image_path):
     # clear the folder 
     folder = PATH_PERSON
     for filename in os.listdir(folder):
+        if filename == '.gitkeep':
+            continue  # Skip deleting .gitkeep file
         file_path = os.path.join(folder, filename)
         try:
             if os.path.isfile(file_path) or os.path.islink(file_path):
                 os.unlink(file_path)
-
         except Exception as e:
             print('Failed to delete %s. Reason: %s' % (file_path, e))
+
 
     for i in range(len(colorize_list_of_masks)):
         image_path = image_path.split('.')[0]
@@ -67,6 +73,7 @@ def second_step(colorize_list_of_masks, image_path):
         cv2.imwrite(name, img)
 
 
+
 def step3(img):
     text_prompt = ['cap', 'shirt', 'sunglasses', 'shoe', 'sock', 'backpack', 'sticks', 'number', 'trousers']
     annotated_image, segmented_image, detections, phrases = detect_and_segment(img, text_prompt)
@@ -76,7 +83,6 @@ def step3(img):
     for i in range(len(detections.mask)):
         
         if  phrases[i] == 'cap' or phrases[i] == 'shirt' or phrases[i] == 'trousers':
-            print (phrases[i])
             filtered_masks.append(colorize_mask(binarize_mask(detections.mask[i]), img))
             filtered_phrases.append(phrases[i])
     #tab_names , average_colors_hexa=determine_color_v2(colorized_masks,phrases)
@@ -94,35 +100,59 @@ def step5 (tab_names , average_colors_hexa, phrases ,croped_bib, text_prompt, p_
     dict_color = {}
     for name , color in zip (tab_names , average_colors_hexa ):
         detected_color = color_or_grayscale(color)
-        print(name ,": ", color_or_grayscale(color),color)
+        #print(name ,": ", color_or_grayscale(color),color)
         dict_color[name] = detected_color
     if croped_bib is not None:
         detected_objects = detect_objects(croped_bib, device=DEVICE)
         bibs= concat_bib_numbers(extract_bib_numbers(sort_bib_numbers(detected_objects)))
 
         # parcourir les phrases et le dictionnaire
-        dict = generate_json_person (text_prompt)
-        print (dict)
-        dict['person'] = p_name
-        for phrase in phrases:
-            if phrase in dict:
-                if phrase in ['cap', 'shirt', 'trousers']:
-                    print (dict_color[phrase])
-                    dict[phrase]['color'] = dict_color[phrase]
-                    dict[phrase]['detected'] = True
-                elif phrase == 'number':
-                    dict[phrase]['detected'] = True
-                    dict[phrase]['numbers'] = bibs
-                else:
-                    dict[phrase]['detected'] = True
-    return dict
+    result_dict = generate_json_person(text_prompt)
+    result_dict['person'] = p_name
 
-def part2():
+    for phrase in phrases:
+        if phrase in result_dict:
+            if phrase in ['cap', 'shirt', 'trousers']:
+                #print(dict_color[phrase])
+                result_dict[phrase]['color'] = dict_color[phrase]
+                result_dict[phrase]['detected'] = True
+            elif phrase == 'number':
+                result_dict[phrase]['detected'] = True
+                result_dict[phrase]['numbers'] = bibs
+            else:
+                result_dict[phrase]['detected'] = True
+
+    return result_dict
+
+def part1(img_path):
+    img = cv2.imread(img_path)
+    colorized_list_of_masks = first_step(img)
+    second_step(colorized_list_of_masks, img_path)
+
+def part2(img_original_name :str):
     # for each image in the folder PATH_PERSON
+    dicts = []
     for filename in os.listdir(PATH_PERSON):
-        img = cv2.imread(PATH_PERSON+filename)
-        p_name = filename.split('.')[0]
+        # verifier le format de nommage de l'image 
+        if filename.startswith('p_'):
+            img = cv2.imread(PATH_PERSON+filename)
+            p_name = filename.split('.')[0]
+            tab_names , average_colors_hexa,detections, phrases, annotated_image,segmented_image,croped_bib , text_prompt= step3(img)
+            n_dict = step5(tab_names , average_colors_hexa, phrases ,croped_bib, text_prompt, p_name)
+            
+            n_dict["imgName"] = img_original_name
+            print(n_dict)
+            dicts.append(n_dict)
+    json_result = []
+    for d in dicts:
+        json_result.append(json.dumps(d))
+    return json_result
 
-        print(dict)
 
-part2()
+def all_steps(img_path):
+    img = cv2.imread(img_path)
+    colorized_list_of_masks = first_step(img)
+    second_step(colorized_list_of_masks, img_path)
+    img_original_name = img_path.split('/')[-1]
+    json_result = part2(img_original_name)
+    return json_result
